refactor(router): split the route url helpers out of the router state
