+++ packages/waku/src/router/client-utils/router-state.ts
@@ -1,49 +1,11 @@
-import {
-  unstable_addBase as addBase,
-  unstable_isImmutableElement as isImmutableElement,
-  unstable_removeBase as removeBase,
-} from '../../minimal/client.js';
-import { pathnameToRoutePath } from '../isomorphic-utils/route-path.js';
+import { unstable_isImmutableElement as isImmutableElement } from '../../minimal/client.js';
 import type { RouteProps } from '../isomorphic-utils/route-path.js';
 import {
   getRouteFromElements,
   getServerRedirect,
   isMetaKey,
 } from './elements-meta.js';
-
-export const pathnameToCurrentRoutePath = (pathname: string) =>
-  pathnameToRoutePath(
-    removeBase(pathname, import.meta.env.WAKU_CONFIG_BASE_PATH),
-  );
-
-export const parseRoute = (url: URL): RouteProps => {
-  const { pathname, searchParams, hash } = url;
-  return {
-    path: pathnameToCurrentRoutePath(pathname),
-    query: searchParams.toString(),
-    hash,
-  };
-};
-
-export const getRouteUrl = (route: RouteProps): URL => {
-  const nextUrl = new URL(window.location.href);
-  nextUrl.pathname = addBase(route.path, import.meta.env.WAKU_CONFIG_BASE_PATH);
-  nextUrl.search = route.query;
-  nextUrl.hash = route.hash;
-  return nextUrl;
-};
-
-export const isSameRoute = (next: RouteProps, prev: RouteProps) =>
-  next.path === prev.path &&
-  next.query === prev.query &&
-  next.hash === prev.hash;
-
-export const parseRedirectUrl = (location: string, base: string | URL) => {
-  const url = new URL(location, base);
-  return url.protocol === 'http:' || url.protocol === 'https:'
-    ? url
-    : undefined;
-};
+import { getRouteUrl } from './route-url.js';
 
 // the client owned router state; the server's ROUTE_ID owns the path
 export const ROUTER_STATE_ID = Symbol('waku-router-state');
